WIP: reorganize UI
